Improved handling of where games results are stored
Copies existing results to a backup folder, and creates a new folder to store per game results.
Added a lock to main.cpp. Probably not necessary, as I don't think code could be re-ordered so as to start processing threads before creating list of starting boards, but put it in just in case.
Games are now numbered in order of when they start, not end.