Added python script that alternates between playing games and NN learning
It uses docker-compose to instantiate the same number of docker containers as game threads, to do inference. The regex string is to ensure that only that number of ports are available. If a larger port range is given, docker-compose does not always start at the beginning, but assigns ports to containers semi-randomly, in which case you do not know which ports have been used (without querying on the command line).